+++ tests/data/google/returns/staticmethod.py
@@ -239,3 +239,20 @@ class MyClass:
                 return False
 
         print(2)
+
+    @staticmethod
+    def zipLists1(
+            list1: List[Any],
+            list2: List[Any],
+    ) -> Iterator[Tuple[Any, Any]]:
+        """
+        Zip 2 lists.
+
+        Args:
+            list1 (List[Any]) : The first list
+            list2 (List[Any]) : The second list
+
+        Returns:
+            Iterator[Tuple[Any, int]]: The zipped result
+        """
+        return zip(list1, list2)