+++ tests/data/numpy/returns/staticmethod.py
@@ -279,3 +279,25 @@ class MyClass:
                 return False
 
         print(2)
+
+    @staticmethod
+    def zipLists1(
+            list1: List[Any],
+            list2: List[Any],
+    ) -> Iterator[Tuple[Any, Any]]:
+        """
+        Zip 2 lists.
+
+        Parameters
+        ----------
+        list1 : List[Any]
+            The first list
+        list2 : List[Any]
+            The second list
+
+        Returns
+        -------
+        Iterator[Tuple[Any, int]]
+            The zipped result
+        """
+        return zip(list1, list2)