+++ tests/data/sphinx/returns/function.py
@@ -233,3 +233,20 @@ def func102(arg0: float):
             return False
 
     print(2)
+
+
+def zipLists1(
+        list1: List[Any],
+        list2: List[Any],
+) -> Iterator[Tuple[Any, Any]]:
+    """
+    Zip 2 lists.
+
+    :param list1: The first list
+    :type list1: List[Any]
+    :param list2: The second list
+    :type list2: List[Any]
+    :return: The zipped list
+    :rtype: Iterator[Tuple[Any, int]]
+    """
+    return zip(list1, list2)